added invitation via mail url

+++ app/Http/Controllers/UserProjectController.php
@@ -2,11 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Mail\InvitationEmail;
 use App\Models\Project;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Str;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
@@ -24,9 +27,11 @@ class UserProjectController extends Controller
         $userName = \request()->get('email');
 
         $user = User::where('email', $userName)->first();
+        $user->setRememberToken(Str::random(60));
+        $user->save();
 
-        if ($user && !$project->isProjectMember($user)) {
-            $project->users()->attach($user->id);
+        if ($user && !$project->isProjectMember($user) && $user->id!== $project->creator_id) {
+            Mail::to($userName)->send(new InvitationEmail($user->remember_token, $project));
             return redirect()->route('projects.index');
         } else {
             return redirect()->route('projects.index')->with('error', "User is already member of project or Invalid username");
@@ -45,7 +50,7 @@ class UserProjectController extends Controller
 
         $user = User::where('email', $userName)->first();
 
-        if ($user && $project->isProjectMember($user)) {
+        if ($user && $project->isProjectMember($user) && $user->id!== $project->creator_id) {
             $project->users()->detach($user->id);
             return redirect()->route('projects.index');
         } else {
@@ -77,5 +82,23 @@ class UserProjectController extends Controller
         } else {
             return redirect()->route('projects.index')->with('error', "User is not member of project or Invalid name");
         }
+    }
+
+    /**
+     * @param string $remember_token
+     * @param User $user
+     * @param Project $project
+     * @return RedirectResponse
+     */
+    public function attachMemberToProject(string $remember_token, Project $project): RedirectResponse
+    {
+        if ($remember_token === Auth::user()->getRememberToken()){
+            if (!$project->isProjectMember(Auth::user()) && Auth::id() !== $project->creator_id){
+                $project->users()->attach(Auth::id());
+                Auth::user()->setRememberToken(null);
+                Auth::user()->save();
+            }
+        }
+        return redirect()->route('projects.index');
     }
 }
